Project rename and docs updates
Rename IdentityServerTestClient -> IdentityTestClient
Update docs on IdentityTestClient and general auth flows
Also some solution fix-ups from previously detleted solutions
