+++ src/TestClients/IdentityTestClient/CommandExtensions.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace IdentityServerTestClient
+namespace IdentityTestClient
 {
     static class CommandExtensions
     {
